# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/message": {
            "get": {
                "description": "Get message content, phone, sent-status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "List messages",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/message.Dto"
                            }
                        }
                    }
                }
            }
        },
        "/start": {
            "post": {
                "tags": [
                    "control"
                ],
                "summary": "Start sender",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "already running",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/stop": {
            "post": {
                "tags": [
                    "control"
                ],
                "summary": "Stop sender",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "not running",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "message.Dto": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdDate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isSent": {
                    "type": "boolean"
                },
                "phone": {
                    "type": "string"
                }
            }
        }
    }
}